+++ extensions/gh-review-raycast/eslint.config.js
@@ -1,2 +1,0 @@
-/** @type {import('eslint').Linter.Config[]} */
-module.exports = require("@raycast/eslint-config");